requirements.txt
alembic==1.15.2
bcrypt==4.3.0
cryptography==44.0.2
fastapi==0.115.12
fastapi-slim==0.115.12
passlib==1.7.4
pydantic==2.11.3
pymysql==1.1.1
python-jose==3.4.0
sqlalchemy==2.0.40
uvicorn==0.34.2
